// src/routing/AppRoute.enum.d.ts
export declare enum AppRoute {
    Login = "/login",
    SignUp = "/sign-up",
    Claim = "/claim",
    Employees = "/employees",
    EmployeeDetails = "/employees/:employeeId",
    Transaction = "/transaction",
    Support = "/support",
    Billing = "/billing",
    ActivityLog = "/activity-log",
    Dashboard = "/dashboard",
    LeaderBoard = "/leaderboard",
    Consultations = "/consultations",
    Patients = "/patients",
    PatientDetails = "/patients/:leadId"
}